pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "castle-wyvern"
version = "0.1.0"
description = "A decentralized multi-agent AI infrastructure with Gargoyles theme"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Chris Tate", email = "blessedwon88@gmail.com"}
]
keywords = ["ai", "multi-agent", "gargoyles", "local-ai", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
local = [
    "ollama>=0.1.0",
]
rag = [
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
all = [
    "ollama>=0.1.0",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/GenieWeenie/castle-wyvern"
Repository = "https://github.com/GenieWeenie/castle-wyvern"
Issues = "https://github.com/GenieWeenie/castle-wyvern/issues"

[project.scripts]
wyvern = "clan_wyvern:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["eyrie*", "grimoorum*", "bmad*", "clan_members*", "spells*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
explicit_package_bases = true
no_implicit_optional = false

[[tool.mypy.overrides]]
module = ["PyPDF2", "docx"]
ignore_missing_imports = true